Review and Retrospective document
We completed the majority of the document on April 4, 2018 when we had our R&R meeting. The selfie was pushed on that day. After talking to the professor on April 9, we learned that we had to push the document. That explains why we are pushing it on April 9. We also learned that there were some sections we needed to fix/add, so we added those. The full edit history can be found here: https://docs.google.com/document/d/17UFWhk8uhC3y9Zvdao-Na9aDSL6LK8WCjtPW9UP0bCU/edit?usp=sharing